+++ CHANGELOG.md
@@ -17,6 +17,8 @@ FEATURES:
  * `kv rollback` Command: You can now use `vault kv rollback` to roll a KVv2
    path back to a previous non-deleted/non-destroyed version. The previous
    version becomes the next/newest version for the path.
+ * Token Bound CIDRs in AppRole: You can now add CIDRs to which a token
+   generated from AppRole will be bound.
 
 IMPROVEMENTS:
 
@@ -42,6 +44,7 @@ IMPROVEMENTS:
 BUG FIXES:
 
  * auth/approle: Fix panic due to metadata being nil [GH-4719]
+ * auth/aws: Fix delete path for tidy operations [GH-4799]
  * core: Optimizations to remove some speed regressions due to the
    security-related changes in 0.10.2
  * secrets/database: Fix default MySQL root rotation statement [GH-4748]